+++ src/markdown2mindmap/core.clj
@@ -15,15 +15,19 @@
 (defn usage [options-summary]
   (->> ["Converts Markdown files to Mind maps."
         ""
-        "Usage: markdown2mindmap [options] <action> <input-file> <output-dir>"
+        "Usage: markdown2mindmap [options] convert <input-file> <output-dir>"
+        "       markdown2mindmap list-all-fonts [output-file]"
         ""
         "Options:"
         options-summary
         ""
         "Actions:"
-        "  convert    Standard convertion."
+        "  convert         Converts markdown file to mindmap."
+        "  list-all-fonts  Creates an SVG image listing all fonts available on the system."
+        "                  output-file defaults to ./all-fonts.svg"
         ""
-        "Example: clojure -M:run-m convert test-resources/input-07.md ."]
+        "Examples: clojure -M:run-m convert test-resources/input-07.md ."
+        "          clojure -M:run-m list-all-fonts"]
        (str/join \newline)))
 
 (defn error-msg [errors]
@@ -39,7 +43,9 @@
   should exit (with a error message, and optional ok status), or a map
   indicating the action the program should take and the options provided."
   [args]
-  (let [{:keys [options arguments errors summary]} (parse-opts args cli-options)]
+  (let [{:keys [options arguments errors summary]} (parse-opts args cli-options)
+        nbargs (count arguments)
+        [action & restargs] arguments]
     (cond
       ;; help => exit OK with usage summary
       (:help options)
@@ -50,9 +56,12 @@
       {:exit-message (error-msg errors)}
 
       ;; custom validation on arguments
-      (and (= 3 (count arguments))
-           (#{"convert"} (first arguments)))
-      {:action (first arguments) :arguments (rest arguments) :options options}
+      (or
+       (and (<= nbargs 2)
+            (#{"list-all-fonts"} action))
+       (and (= 3 nbargs)
+            (#{"convert"} action)))
+      {:action action :arguments restargs :options options}
 
       ;; failed custom validation => exit with usage summary
       :else
@@ -60,12 +69,14 @@
 
 (defn -main [& args]
   (let [{:keys [action options arguments exit-message ok?]} (validate-args args)
-        [input-file output-directory] arguments]
-    (prn input-file output-directory (:type options))
+        [input-file output-directory] arguments
+        [output-file] arguments]
     (if exit-message
       (exit (if ok? 0 1) exit-message)
       (case action
-        "convert" (m2mtransform/md->mindmap input-file
-                                            output-directory
-                                            (:type options)))))
+        "convert"        (m2mtransform/md->mindmap input-file
+                                                   output-directory
+                                                   (:type options))
+        "list-all-fonts" (m2mtransform/list-all-fonts
+                          (or output-file "./all-fonts.svg")))))
   (exit 0 ":ok"))